refactor(bench-support): scaffold generator submodules and test infrastructure
Extract CsvPayload and JSON generators into generators/ submodules.
Add seed parameter to all generator signatures (breaking change).
Update 9 call sites to pass seed: 42. Add cache.rs and XML/fixed-width stubs.
