+++ crates/clinker-bench-support/src/cache.rs
@@ -1,0 +1,3 @@
+//! Benchmark data file cache with `.meta.json` sidecars.
+//!
+//! Stub — implementation in Task 2.4.
