+++ crates/clinker-bench-support/src/generators/fixed_width.rs
@@ -1,0 +1,3 @@
+//! Deterministic fixed-width payload generator for benchmarks.
+//!
+//! Stub — implementation in Task 2.2.
