+++ crates/clinker-bench-support/src/generators/mod.rs
@@ -1,0 +1,6 @@
+//! Format-specific data generators for benchmarks.
+
+pub mod csv;
+pub mod fixed_width;
+pub mod json;
+pub mod xml;
