+++ crates/clinker-bench-support/src/generators/xml.rs
@@ -1,0 +1,3 @@
+//! Deterministic XML payload generator for benchmarks.
+//!
+//! Stub — implementation in Task 2.1.
