+++ crates/clinker-core/benches/pipeline.rs
@@ -82,7 +82,7 @@ nodes:
     let params = test_params();
 
     for count in [SMALL, MEDIUM] {
-        let csv_bytes = CsvPayload::generate(count, 5, 16);
+        let csv_bytes = CsvPayload::generate(count, 5, 16, 42);
 
         group.throughput(Throughput::Elements(count as u64));
         group.bench_with_input(BenchmarkId::from_parameter(count), &count, |b, _| {
@@ -163,7 +163,7 @@ nodes:
     let params = test_params();
 
     for count in [SMALL, MEDIUM] {
-        let csv_bytes = CsvPayload::generate(count, 5, 16);
+        let csv_bytes = CsvPayload::generate(count, 5, 16, 42);
 
         group.throughput(Throughput::Elements(count as u64));
         group.bench_with_input(BenchmarkId::from_parameter(count), &count, |b, _| {
@@ -258,7 +258,7 @@ nodes:
     let params = test_params();
 
     for count in [SMALL, MEDIUM] {
-        let csv_bytes = CsvPayload::generate(count, 5, 16);
+        let csv_bytes = CsvPayload::generate(count, 5, 16, 42);
 
         group.throughput(Throughput::Elements(count as u64));
         group.bench_with_input(BenchmarkId::from_parameter(count), &count, |b, _| {
@@ -350,7 +350,7 @@ nodes:
     let params = test_params();
 
     for count in [SMALL, MEDIUM] {
-        let csv_bytes = CsvPayload::generate(count, 5, 16);
+        let csv_bytes = CsvPayload::generate(count, 5, 16, 42);
 
         group.throughput(Throughput::Elements(count as u64));
         group.bench_with_input(BenchmarkId::from_parameter(count), &count, |b, _| {
